Move validator interfaces to virsat.model project - (Task #566)
