+++ de.dlr.sc.virsat.build/META-INF/MANIFEST.MF
@@ -17,6 +17,5 @@ Export-Package: de.dlr.sc.virsat.build.inheritance,
  de.dlr.sc.virsat.build.marker.util,
  de.dlr.sc.virsat.build.validator,
  de.dlr.sc.virsat.build.validator.core,
- de.dlr.sc.virsat.build.validator.external,
  de.dlr.sc.virsat.inheritance
 Automatic-Module-Name: de.dlr.sc.virsat.build
