+++ de.dlr.sc.virsat.model/META-INF/MANIFEST.MF
@@ -56,6 +56,7 @@ Export-Package: de.dlr.sc.virsat.model.dvlm,
  de.dlr.sc.virsat.model.dvlm.units.util,
  de.dlr.sc.virsat.model.dvlm.util,
  de.dlr.sc.virsat.model.dvlm.util.rules,
+ de.dlr.sc.virsat.model.dvlm.validator,
  de.dlr.sc.virsat.model.ecore,
  de.dlr.sc.virsat.model.ecore.xmi.impl
 Automatic-Module-Name: de.dlr.sc.virsat.model
