+++ de.dlr.sc.virsat.model/src/de/dlr/sc/virsat/model/dvlm/validator/IRepositoryValidator.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.build.validator.external;
+package de.dlr.sc.virsat.model.dvlm.validator;
 
 import de.dlr.sc.virsat.model.dvlm.Repository;
